Update to u-boot v2019.04
Update to u-boot v2019.04 and refresh the u-boot patch as needed. The
reason for the update is fixing a bug for which loading some device tree
overlays from the first stage bootloader causes a u-boot crash. For
instance, adding the line

dtoverlay=pi3-disable-bt

to config.txt provokes this. This does not happen anymore in v2019.04.