Split the boot configuration
The main config.txt needs to call different u-boot binaries for
different pi revisions (mostly to support the pi4). In addition, it's
desirable to have config.txt essentially static with modifiable
includes. This commit moves to this model and also includes some extra
configuration examples.

In concert with this change, the u-boot-rpi package will be updated to
include a postinst which will migrate existing installations to this
style of configuration.

After testing it appears that on older firmwares conditional [sections]
aren't recognized. Hence the bootloader will read the [pi3] section and
then carry on to read the [pi4] section! Naturally, the pi4 u-boot isn't
optimized for use on the pi3, but more importantly the pi4 u-boot won't
exist at all on images built with older u-boot-rpi packages.

As a workaround, specify kernels in an unusual order (4 first, then 2,
then 3) so that later settings override earlier ones. This results in
the correct kernel (uboot) selection with both old and new firmwares
across platforms.