Use boot script from flash-kernel
Supports both archs and compression on each (which the bootscript in
this repo doesn't).

NOTE: The RPi bootscript was named bootscr.rpi3 in bionic (along with a
ubootenv for the pi2); in subsequent releases this simply became a
generic bootscr.rpi. This explains the "*" wildcard on the source of
mkimage.